move similarity query into LiteDB
+ change BucketVariance setting to be a byte
+ change Buckets into ushort[]
+ remove unused CountBits() method
+ remove old SimilarityQuery() method
+ change buckets for prefilter to be static, clear before using
+ remove SimilarityQueryResult class
+ simplify similarity query by adding SIMILARITY() and SIMILARITY_AVERAGED() to LiteDB.dll
    - no difference in speed (2x faster if leave off query.Count() call)
    - should use significantly less memory (no longer necessary to temporarily store an array of SimilarityQueryResult containing the relevant hashes for every image not filtered out by prefilter
    - requires less code overall

+++ src/Core/ImageInfo.cs
@@ -6,7 +6,7 @@ namespace ImageTagger.Core
 {
     public sealed class ImageInfo
     {
-        private static readonly int[] defaultBuckets = new int[4];
+        private static readonly ushort[] defaultBuckets = new ushort[4];
 
         public string Hash { get; set; }
         public string Name { get; set; }
@@ -17,7 +17,7 @@ namespace ImageTagger.Core
         public ulong PerceptualHash { get; set; }
         public ulong ColorHash { get; set; }
 
-        public int[] Buckets { get; set; }
+        public ushort[] Buckets { get; set; }
         public int NumFrames { get; set; }
         public int[] Colors { get; set; }
